refactor(jotai): move sidebar/modal atoms from admin to common scope
