+++ src/components/admin/account/info/AccountInfo.tsx
@@ -1,6 +1,7 @@
 import styled from '@emotion/styled';
 import { colFlex } from '@styles/flexStyles';
-import { adminUserAccountAtom, externalSidebarAtom } from '@jotai/admin/atoms';
+import { adminUserAccountAtom } from '@jotai/admin/atoms';
+import { externalSidebarAtom } from '@jotai/atoms';
 import { useAtomValue, useSetAtom } from 'jotai';
 import { ACCOUNT_INFO, ACCOUNT_MODAL } from '@constants/data/accountData';
 import RegisterAccountInfoContainer from './RegisterAccountInfoContainer';
