+++ src/components/admin/account/register/RegisterAccount.tsx
@@ -4,7 +4,8 @@ import { useEffect, useRef, useState } from 'react';
 import styled from '@emotion/styled';
 import { colFlex, rowFlex } from '@styles/flexStyles';
 import NewAppInput from '@components/common/input/NewAppInput';
-import { adminBanksAtom, externalSidebarAtom } from '@jotai/admin/atoms';
+import { adminBanksAtom } from '@jotai/admin/atoms';
+import { externalSidebarAtom } from '@jotai/atoms';
 import { useAtomValue, useSetAtom } from 'jotai';
 import NewCommonButton from '@components/common/button/NewCommonButton';
 import { RIGHT_SIDEBAR_ACTION } from '@@types/index';
