+++ src/components/admin/order/realtime/modal/order-detail/OrderDetailModal.tsx
@@ -8,7 +8,7 @@ import OrderModalHeaderContents from '@components/admin/order/realtime/modal/ord
 import { createPortal } from 'react-dom';
 import { useEffect } from 'react';
 import useModal from '@hooks/useModal';
-import { orderModalReadOnlyAtom } from '@jotai/admin/atoms';
+import { orderModalReadOnlyAtom } from '@jotai/atoms';
 import { useSetAtom } from 'jotai';
 
 const DetailModalContainer = styled.div``;
