+++ src/pages/admin/table/AdminTableRealtime.tsx
@@ -19,7 +19,8 @@ import { colFlex, rowFlex } from '@styles/flexStyles';
 import { useEffect } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
 import { useAtomValue, useSetAtom } from 'jotai';
-import { adminTablesAtom, adminWorkspaceAtom, externalSidebarAtom } from '@jotai/admin/atoms';
+import { adminTablesAtom, adminWorkspaceAtom } from '@jotai/admin/atoms';
+import { externalSidebarAtom } from '@jotai/atoms';
 import { RIGHT_SIDEBAR_ACTION } from '@@types/index';
 import NewCommonButton from '@components/common/button/NewCommonButton';
 import { RiSettings3Fill } from '@remixicon/react';
